job_board/pyproject.toml
[project]
name = "job-board"
version = "0.1.0"
description = ""
authors = [
    {name = "mukhtar-ahmed",email = "kh.mukhtar60@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
